FAssert can now be used as an if condition.
This was accomplished by moving the guts of the assertion into a function, and simply having the macro call that function with the appropriate arguments. This also meant that the FFailWithOptionalMessageString macro didn’t need to be used (although it is retained for the sake of FFail).

This lets you do convenient guarded assertions, for example in cases where later assertions depend on the former passing and will crash if it does not. Example:

	FooStructure *foo = FooStructureCreateWithMagicalFairyDust();
	if(FAssert(foo != NULL)) {
		FAssert(foo->wasCreatedWithMagicalFairyDust);
	}

Without the if statement, the second assert would cause a crash in cases where the first failed.

This behaviour could be replicated before by performing the test twice (once in the assert, once in the if condition) or by storing the condition in a variable, but this implementation of it is particularly pleasing.